+++ tests/MarkdownTest.php
@@ -16,6 +16,9 @@ namespace Eventum\Test;
 use Eventum\Markdown;
 use Generator;
 
+/**
+ * @group db
+ */
 class MarkdownTest extends TestCase
 {
     /** @var Markdown */
@@ -53,6 +56,7 @@ class MarkdownTest extends TestCase
             'smartpunct',
             'table',
             'tasklist',
+            'userhandle',
         ];
 
         foreach ($testNames as $testName) {